refactor:整理不再使用的程式碼 及更名
