+++ app.js
@@ -10,7 +10,7 @@ require("./utils/conn.js");
 
 var indexRouter = require("./routes/index");
 var usersRouter = require("./routes/users");
-var postsRouter = require("./routes/posts");
+
 
 var newPostRouter = require("./routes/article");
 var app = express();
@@ -30,7 +30,6 @@ app.use(express.static(path.join(__dirname, "public")));
 
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
-app.use("/posts", postsRouter);
 app.use("/article", newPostRouter);
 app.use("/api-doc", swaggerUi.serve, swaggerUi.setup(swaggerFile));
